Implement retention policy for dev in Artifactory
The retention policy is coupled with the build information so we need to
enable the creation of Builds for CS Studio in Artifactory as well.
This should have the nice side effect that you will now have a way to
browse which files have been uploaded to gether as artifacts from the
same build in Jenkins.

JIRA INFRA-462 #action Review